fix: Update to opensafely-jobrunner v2.13.1

+++ opensafely/_vendor/jobrunner/manage_jobs.py
@@ -81,6 +81,9 @@ def start_job(job):
             env["DATABASE_URL"] = config.DATABASE_URLS[job.database_name]
             if config.TEMP_DATABASE_NAME:
                 env["TEMP_DATABASE_NAME"] = config.TEMP_DATABASE_NAME
+            if config.PRESTO_TLS_KEY and config.PRESTO_TLS_CERT:
+                env["PRESTO_TLS_CERT"] = config.PRESTO_TLS_CERT
+                env["PRESTO_TLS_KEY"] = config.PRESTO_TLS_KEY
     # Prepend registry name
     image = action_args[0]
     full_image = f"{config.DOCKER_REGISTRY}/{image}"
@@ -299,9 +302,12 @@ def finalise_job(job):
 
 
 def cleanup_job(job):
-    log.info("Cleaning up container and volume")
-    docker.delete_container(container_name(job))
-    docker.delete_volume(volume_name(job))
+    if config.CLEAN_UP_DOCKER_OBJECTS:
+        log.info("Cleaning up container and volume")
+        docker.delete_container(container_name(job))
+        docker.delete_volume(volume_name(job))
+    else:
+        log.info("Leaving container and volume in place for debugging")
 
 
 def get_container_metadata(job):
